Update PKHeX.Core To 24.03.10 And Version To 1.7.0
MAR10 Day Update

- Replace StringUtils.GetUnicodeString with StringConverter8.GetString
- Fix forgetting to account for the offset in Musical5.Data for the props that is done in commit 24cafd51268f9c31d5e37d5c9352502beba58332 when switching to using PKHeX Gen 5 structures for BWTool

+++ PluginPile.Common/DrawingUtil.cs
@@ -15,7 +15,7 @@ public class DrawingUtil {
     return (Image)getSpriteMethod.Invoke(null, [pkm])!;
   }  
 
-  public static Image GetSprite(ushort species, byte form, int gender, uint formarg, int item, bool isegg, Shiny shiny, EntityContext context = EntityContext.None) {
+  public static Image GetSprite(ushort species, byte form, byte gender, uint formarg, int item, bool isegg, Shiny shiny, EntityContext context = EntityContext.None) {
     MethodInfo getSpriteMethod = SpriteUtilType.GetMethods(BindingFlags.Public | BindingFlags.Static)
       .Single(m => m.Name == "GetSprite" && m.GetParameters().Length == 8);
     return (Image)getSpriteMethod.Invoke(null, [species, form, gender, formarg, item, isegg, shiny, context])!;
